+++ R/get_fcc_dictionary.R
@@ -1,6 +1,6 @@
 #' Display FCC variables and associated descriptions
 #'
-#' Return dictionary for a all datasets. Available dataset are "f477", "nbm_raw" and "nbm_block".
+#' Return dictionary for a all datasets. Available dataset are "f477" (`get_f477`), "nbm_raw" (`get_nbm_county_raw`) and "nbm_block" (`get_nbm_bl`, `get_nbm_county`).
 #'
 #' @param dataset a string matching a dataset, default is "all"
 #'
@@ -11,7 +11,9 @@
 #' @examples
 #' get_fcc_dictionary("nbm_block")
 
-get_fcc_dictionary <- function(dataset = "all") {
+get_fcc_dictionary <- function(dataset = c("all", "f477", "nbm_raw", "nbm_block")) {
+  dataset <- match.arg(dataset)
+
   dict <- cori.data.fcc::fcc_dictionary
   if (dataset == "all") {
     return(dict)